alias maximum to clamp_min
